refactor: change radius type from Double to Int

+++ app/src/main/java/com/github/se/icebreakrr/mock/MockProfiles.kt
@@ -48,7 +48,7 @@ open class MockProfileRepository : ProfilesRepository {
   // Returns a predefined list of profiles for getProfilesInRadius
   override fun getProfilesInRadius(
       center: GeoPoint,
-      radiusInMeters: Double,
+      radiusInMeters: Int,
       onSuccess: (List<Profile>) -> Unit,
       onFailure: (Exception) -> Unit
   ) {
